update node select for filters
